Separated printing functions into databaseTest.py, added functionality to support dumping db tables from command line with a single command
